+++ Lesson_1/4.rb
@@ -1,9 +1,9 @@
 print "Vvedite koef a: "
-a = gets.to_i
+a = gets.to_f
 print "Vvedite koef b: "
-b = gets.to_i
+b = gets.to_f
 print "Vvedite koef c: "
-c = gets.to_i
+c = gets.to_f
 
 D = b**2 - 4*a*c
 
@@ -12,12 +12,12 @@ if D < 0
 end
 
 if D > 0 
-    x1 = (-b + Math.sqrt(D)) / (2*a) 
-    x2 = (-b - Math.sqrt(D)) / (2*a)
-    puts "D=#{D}, x1 = #{x1}, x2 = #{x2}"
+  x1 = (-b + Math.sqrt(D)) / (2*a) 
+  x2 = (-b - Math.sqrt(D)) / (2*a)
+  puts "D=#{D}, x1 = #{x1}, x2 = #{x2}"
 end
 
 if D == 0
-   x = -b/(2*a)
-   puts "D=#{D}, x = #{x}"
+  x = -b/(2*a)
+  puts "D=#{D}, x = #{x}"
 end
